Character.setAcceptableAlignment: Being rich is an influence towards LE
There's a whole bunch of studies showing that as people grow wealthier, they grow more selfish and more invested in preserving the legal structures that gave them their wealth and power. That's LE in a nutshell. Just like culture, race, or religion, it's not saying outright that every rich person *is* LE, but it's an influence pulling you in that direction, just like those other external factors.